✨ Criação de templates de issue e pull request ✨ Adiciona workflow de CI para testes e cobertura de código ✨ Adiciona workflow de release para publicação de novas versões ✨ Cria arquivo CHANGELOG.md para registrar as mudanças do projeto ✨ Cria arquivo CODE_OF_CONDUCT.md para definir o código de conduta do projeto ✨ Cria arquivo CONTRIBUTING.md para guiar as contribuições ao projeto ✨ Cria arquivo SECURITY.md para informar sobre a política de segurança do projeto 🐛 Corrige bug na função RoundToNearestMinutes ✨ Adiciona exemplos de uso na documentação ✨ Adiciona exemplos de uso no arquivo README.md ✨ Adiciona exemplos de uso no arquivo examples/main.go ♻️ Refatora o arquivo examples/main.go para melhor organização ⬆️ Atualiza versão do Go para 1.21 ⬆️ Atualiza licença para LGPL v2.1 🗑️ Remove arquivo IMPLEMENTATION_SUMMARY.md

+++ dateutils/additional_utility_functions.go
@@ -116,7 +116,9 @@ func RoundToNearestMinutes(t time.Time, nearestTo int) time.Time {
 	remainder := minutes % nearestTo
 
 	var adjustedMinutes int
-	if remainder < nearestTo/2 {
+	// Use float64 comparison to avoid integer division truncation.
+	// e.g. for nearestTo=5, threshold should be 2.5, not 2 (int division).
+	if float64(remainder) < float64(nearestTo)/2.0 {
 		// Round down
 		adjustedMinutes = minutes - remainder
 	} else {
